front: InputGroupSNCF: typescript & correct parents
